requirements.txt
opencv-python
ultralytics
google-generativeai
pyttsx3
speechrecognition
pyserial
torch
